Add native/web apps dfns, expand on native/web bridge
This adds definitions for "native application" and "web application" to clarify
the context under which WebViews are being used. The "access to the web content"
concept was also completed to clarify that it allows native applications to
create a bridge between the native world and the web world and expose native
APIs to the Web content in particular.

The table now refers to this "access to the web content" concept directly.